Cleaned up system stuff significantly.
