+++ electrical/SandboxNucleoBoard.c
@@ -1,6 +1,6 @@
 // Welcome to the sandbox! Please read the comments at the end of the file.
 
-#include "defs.h"
+#include "system.h"
 #include "uxart.c"
